fix(cli): keep SDK bundle off the Ink TUI graph
App create imported onboarding report which pulled runtime.tsx, so the SDK build failed on optional react-devtools-core.
